requirements.txt
asyncio~=3.4.3
Telethon~=1.33.1
psycopg~=3.1.14
psycopg-binary~=3.1.13
python-dotenv~=1.0.0
SQLAlchemy~=2.0.23
alembic~=1.13.0
openai~=1.3.7
python-socks~=2.4.3
async_timeout
requests~=2.31.0
aiogram~=3.2.0
aiohttp~=3.9.1